requirements.txt
twitchio
pyyaml
pytz
requests
